+++ data/metric/metrics.go
@@ -28,6 +28,7 @@ var (
 	ErrNonNumeric        = errors.New("non-numeric value for rate/delta")
 	ErrNoStoreToCalcDiff = errors.New("cannot use deltas nor rates without persistent store")
 	ErrTooCloseSamples   = errors.New("samples too close in time, skipping")
+	ErrNegativeDiff      = errors.New("source was reset, skipping")
 	ErrOverrideSetAttrs  = errors.New("cannot overwrite metric-set attributes")
 	ErrDeltaWithNoAttrs  = errors.New("delta/rate metrics should be attached to an attribute identified metric-set")
 )
@@ -80,7 +81,7 @@ func (ms *Set) SetMetric(name string, value interface{}, sourceType SourceType) 
 
 	// Only sample metrics of numeric type
 	switch sourceType {
-	case RATE, DELTA:
+	case RATE, DELTA, PRATE, PDELTA:
 		if len(ms.nsAttributes) == 0 {
 			err = ErrDeltaWithNoAttrs
 			return
@@ -163,6 +164,12 @@ func (ms *Set) elapsedDifference(name string, absolute interface{}, sourceType S
 	}
 
 	elapsed = newValue - oldValue
+
+	if elapsed < 0 && sourceType.IsPositive() {
+		err = ErrNegativeDiff
+		return
+	}
+
 	if sourceType == RATE {
 		elapsed = elapsed / float64(duration)
 	}