+++ data/metric/source_type.go
@@ -21,6 +21,10 @@ const (
 	DELTA SourceType = iota
 	// ATTRIBUTE is any string value
 	ATTRIBUTE SourceType = iota
+	// PRATE is a version of RATE that only allows positive values.
+	PRATE SourceType = iota
+	// PDELTA is a version of DELTA that only allows positive values.
+	PDELTA SourceType = iota
 )
 
 // SourcesTypeToName metric sources list mapping its type to readable name.
@@ -36,16 +40,24 @@ var SourcesNameToType = map[string]SourceType{
 	"gauge":     GAUGE,
 	"rate":      RATE,
 	"delta":     DELTA,
+	"prate":     PRATE,
+	"pdelta":    PDELTA,
 	"attribute": ATTRIBUTE,
 }
 
 // String fulfills stringer interface, returning empty string on invalid source types.
-func (t *SourceType) String() string {
-	if s, ok := SourcesTypeToName[*t]; ok {
+func (t SourceType) String() string {
+	if s, ok := SourcesTypeToName[t]; ok {
 		return s
 	}
 
 	return ""
+}
+
+// IsPositive checks that the `SourceType` belongs to the positive only
+// list of `SourceType`s
+func (t SourceType) IsPositive() bool {
+	return t == PRATE || t == PDELTA
 }
 
 // SourceTypeForName does a case insensitive conversion from a string to a SourceType.