+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.2.0
+
+### Added
+
+- Adds the positive only version of `RATE` and `DELTA`, named `PRATE` and
+  `PDELTA` respectively.
+
 ## 3.1.5
 
 ### Fixed